[Refactor]: OAuthAuthenticationProvider를 유틸 클래스로 변경
본문

[#?]
