+++ src/main/java/project/common/config/OAuth2Config.java
@@ -1,6 +1,6 @@
 package project.common.config;
 
-import project.common.security.oauth.OAuthAuthenticationProvider;
+import project.common.util.Oauth2UrlUtil;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
@@ -10,8 +10,8 @@ import org.springframework.context.annotation.Configuration;
 public class OAuth2Config {
     @Bean
     @ConfigurationProperties(prefix = "oauth2")
-    OAuthAuthenticationProvider oAuthAuthenticationProvider(){
-        return new OAuthAuthenticationProvider();
+    Oauth2UrlUtil oAuthAuthenticationProvider(){
+        return new Oauth2UrlUtil();
     }
 
 }
